+++ docs/pages/release_notes.md
@@ -20,6 +20,8 @@ This is a {{ site.pmd.release_type }} release.
 * java
   * [#5293](https://github.com/pmd/pmd/issues/5293): \[java] Deadlock when executing PMD in multiple threads
   * [#5324](https://github.com/pmd/pmd/issues/5324): \[java] Issue with type inference of nested lambdas
+* java-bestpractices
+  * [#5318](https://github.com/pmd/pmd/issues/5318): \[java] PreserveStackTraceRule: false-positive on Pattern Matching with instanceof
 
 ### 🚨 API Changes